Updated logging + Twitch data checking
Updated the logging to now include the date (although not set to my time zone yet). I found that there was an error on if(data[0]) because when there's no data, it's undefined and tries to check for the 0th spot of an undefined JSON object. Changed that to check for if data[0] is undefined or not, and that resolves the error being put into the logs.

Also, not included, but I had to update my twitch_token because it had changed.