Birthday Event Skeleton
Adding a skeleton for a command to use before/during my birthday stream when I want to have events for it. This particular year, I'm going to have viewers guess whether or not I beat a challenge I'll be attempting live.

Using MySQL to do this, so got up basics for it for now because something's going on with my MySQL instance.