bday.js error catching
When an error from the SQL Query happens, it's logged and code continues to run; this can cause the bot to crash if rows is undefined. Now, it should stop when an error is hit, let the user know there was an error, and stop the rest of the code from running so the bot doesn't crash.