Checking for Twitch API data
Now checks if the data being returned from the Twitch API call is undefined. If it is, then there was an issue connecting to Twitch. Logging the error, putting a message in a bot log channel so I can be notified, and returns because it can't do anything with no response body at all.