Bringing in all new code changes to bot.js
bot.js now imports all files (may clean this up later) and utilizes all new utility functions.

Big change is to api variable assignment. I now check to see if the token has been changed, and if it has, re-assign the variables so that they can update automatically. With this, I no longer need to keep tabs on the expiration of a token -- it'll do it all on its own.